pyproject.toml
[project]
name = "ragdoll-comfy-nodes"
version = "0.1.0"
description = "ComfyUI nodes for ragdoll"
authors = [{ name = "the-dream-machine" }]
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "boto3==1.40.13",
    "boto3-stubs==1.40.13",
    "cuid2==2.0.1",
    "numpy==2.3.2",
    "pillow==11.3.0",
    "torch==2.8.0",
]


[tool.uv]
add-bounds = "exact"
